move HasHttpManager from http-conduit to http-client
